experiment/testResults/requirements.txt
requests
scikit-learn
umap-learn
matplotlib
requests
beautifulsoup4
numpy
umap-learn
matplotlib
nltk==3.8.1
dotenv
pandas
seaborn